refactor(authentication): Extract PasswordGeneratorService to use an interface
