+++ backend/src/main/java/com/project/G1_T3/admin/service/AdminService.java
@@ -9,7 +9,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
-import com.project.G1_T3.authentication.service.PasswordGeneratorService;
+import com.project.G1_T3.authentication.service.PasswordGeneratorServiceImpl;
 import com.project.G1_T3.user.model.User;
 import com.project.G1_T3.user.model.UserDTO;
 import com.project.G1_T3.user.model.UserRole;
@@ -28,7 +28,7 @@ public class AdminService {
     private UserService userService;
 
     @Autowired
-    private PasswordGeneratorService passwordGeneratorService;
+    private PasswordGeneratorServiceImpl passwordGeneratorService;
 
     @Autowired
     private UserRepository userRepository;
